rename all how_to files to how-to
