+++ src/app/help/how-to/how-to.component.spec.ts
