+++ src/app/help/how-to/how-to.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-help',
-  templateUrl: './how_to.component.html',
-  styleUrls: ['./how_to.component.scss']
+  templateUrl: './how-to.component.html',
+  styleUrls: ['./how-to.component.scss']
 })
 export class HowToComponent implements OnInit {
